GODRIVER-1646: Unmarshal errors should return multiple errors when possible
Started implementing a MultiDecodeError type that allows to returns all
decoding errors wrapped in a single error.

And also GODRIVER-1645: Type Mismatch Errors during decode should use a custom type
To be able to compare the type that mismatched during the decoding.

Added a toggle (SetIgnoreDecodingError) on the RegistryBuilder set to true by default in mgoCompat
to log and ignore the errors, allowing to have mgo like bson unmarshal error skipping while still
logging all errors to fix productions databases when needed.

TODO:
Return the decode errors to the library user rather than printing it to the console in mgoCompat.
Fix all tests : skip in bsonrwtest.go does not works the same as in value_reader.go

Signed-off-by: Julien Spadoni <julien@beezeelinx.com>